requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
aniso8601==9.0.1
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
beautifulsoup4==4.12.3
black==19.10b0
boto3==1.34.44
botocore==1.34.44
braceexpand==0.1.7
cdifflib==1.2.6
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
ctc-decoders==1.1
cycler==0.12.1
Cython==3.0.8
cytoolz==0.12.3
datasets==2.17.0
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.8.1
einops==0.7.0
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fastapi==0.111.0
fastapi-cli==0.0.3
fasteners==0.19
fasttext==0.9.2
fasttext-wheel==0.9.2
filelock @ file:///croot/filelock_1700591183607/work
flashlight-text @ file:///lfs01/workdirs/hlwn038u1/NeMo/nemo/decoders/text/dist/flashlight_text-0.0.7%2B4a2b9b0.d20240416-cp310-cp310-linux_x86_64.whl#sha256=d45ad39ff32bef5a19a592538a3478c19692fa7d7af23ae432d2a85f2617072c
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
g2p-en==2.1.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.20.3
hydra-core==1.3.2
hypothesis==6.99.13
idna @ file:///croot/idna_1666125576474/work
ijson==3.2.3
imagesize==1.4.1
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
ipython==8.21.0
ipywidgets==8.1.2
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2 @ file:///croot/jinja2_1706733616596/work
jiwer==2.5.2
jmespath==1.0.1
joblib==1.3.2
jupyterlab_widgets==3.0.10
kaldi-python-io==1.2.2
kaldiio==2.18.0
kenlm==0.2.0
kiwisolver==1.4.5
kornia==0.7.1
latexcodec==2.0.1
lazy_loader==0.3
Levenshtein==0.22.0
lhotse==1.21.0
librosa==0.10.1
lightning-utilities==0.10.1
lilcom==1.7
llvmlite==0.42.0
loguru==0.7.2
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
megatron-core==0.3.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
nemo-text-processing==0.1.8rc0
nemo-toolkit==1.21.0
networkx @ file:///croot/networkx_1690561992265/work
nltk==3.8.1
numba==0.59.0
numcodecs==0.12.1
numpy==1.23.5
omegaconf==2.3.0
onnx==1.15.0
OpenCC==1.1.6
orjson==3.10.3
packaging==23.2
pandas==2.2.0
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow @ file:///croot/pillow_1707233021655/work
plac==1.4.2
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.0
portalocker==2.8.2
progress==1.6
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
PyArabic==0.6.15
pyarrow==15.0.0
pyarrow-hotfix==0.6
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.21
pyctcdecode==0.5.0
pydantic==1.10.14
pydub==0.25.1
Pygments==2.17.2
pygtrie==2.5.0
pyngrok==7.1.6
pynini==2.1.5
pyparsing==3.1.1
pypinyin==0.50.0
pypinyin-dict==0.7.0
PySocks==1.7.1
pytest==8.0.1
pytest-runner==6.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.0.7
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
rapidfuzz==2.13.7
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
rich==13.7.0
rouge-score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sacrebleu==2.4.0
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
sentry-sdk==1.40.4
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.7
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
starlette==0.37.2
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.1
texterrors==0.4.4
threadpoolctl==3.3.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchmetrics==1.3.1
torchvision==0.17.0
tqdm==4.66.2
traitlets==5.14.1
transformers==4.33.3
triton==2.2.0
typed-ast==1.5.5
typer==0.12.3
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2024.1
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.16.3
watchfiles==0.21.0
wcwidth==0.2.13
webdataset==0.1.62
websockets==12.0
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtokentome==1.0.6
zarr==2.17.0
